+++ doc/release/v3_1_0.md
@@ -17,7 +17,38 @@ New Features
 
 #### YARP_DEV
 
-* File ControlCalibrationImpl.cpp renamed to IControlCalibrationImpl.cpp
+* The following files have been ranamed:
+  src/libYARP_dev/include/yarp/dev/ImplementControlLimits.h (from src/libYARP_dev/include/yarp/dev/IControlLimitsImpl.h)
+  src/libYARP_dev/include/yarp/dev/ImplementControlLimits2.h (from src/libYARP_dev/include/yarp/dev/IControlLimits2Impl.h)
+  src/libYARP_dev/include/yarp/dev/ImplementInteractionMode.h (from src/libYARP_dev/include/yarp/dev/IInteractionModeImpl.h)
+  src/libYARP_dev/include/yarp/dev/ImplementPidControl.h (from src/libYARP_dev/include/yarp/dev/IPidControlImpl.h)
+  src/libYARP_dev/include/yarp/dev/ImplementPositionControl.h (from src/libYARP_dev/include/yarp/dev/IPositionControlImpl.h)
+  src/libYARP_dev/include/yarp/dev/ImplementPositionControl2.h (from src/libYARP_dev/include/yarp/dev/IPositionControl2Impl.h)
+  src/libYARP_dev/include/yarp/dev/ImplementPositionDirect.h (from src/libYARP_dev/include/yarp/dev/IPositionDirectImpl.h)
+  src/libYARP_dev/include/yarp/dev/ImplementVelocityControl.h (from src/libYARP_dev/include/yarp/dev/IVelocityControlImpl.h)
+  src/libYARP_dev/include/yarp/dev/ImplementVelocityControl2.h (from src/libYARP_dev/include/yarp/dev/IVelocityControl2Impl.h)
+  src/libYARP_dev/include/yarp/dev/ImplementVirtualAnalogSensor.h (from src/libYARP_dev/include/yarp/dev/IVirtualAnalogSensorImpl.h)
+  src/libYARP_dev/src/ImplementAmplifierControl.cpp (from src/libYARP_dev/src/IAmplifierControlImpl.cpp)
+  src/libYARP_dev/src/ImplementAxisInfo.cpp (from src/libYARP_dev/src/IAxisInfoImpl.cpp)
+  src/libYARP_dev/src/ImplementControlCalibration.cpp (from src/libYARP_dev/src/IControlCalibrationImpl.cpp)
+  src/libYARP_dev/src/ImplementControlLimits.cpp (from src/libYARP_dev/src/IControlLimitsImpl.cpp)
+  src/libYARP_dev/src/ImplementControlMode.cpp (from src/libYARP_dev/src/ControlModeImpl.cpp)
+  src/libYARP_dev/src/ImplementCurrentControl.cpp (from src/libYARP_dev/src/CurrentControlImpl.cpp)
+  src/libYARP_dev/src/ImplementEncoders.cpp (from src/libYARP_dev/src/IEncodersImpl.cpp)
+  src/libYARP_dev/src/ImplementEncodersTimed.cpp (from src/libYARP_dev/src/IEncodersTimedImpl.cpp)
+  src/libYARP_dev/src/ImplementImpedanceControl.cpp (from src/libYARP_dev/src/ImpedanceControlImpl.cpp)
+  src/libYARP_dev/src/ImplementInteractionMode.cpp (from src/libYARP_dev/src/IInteractionModeImpl.cpp)
+  src/libYARP_dev/src/ImplementMotor.cpp (from src/libYARP_dev/src/IMotorImpl.cpp)
+  src/libYARP_dev/src/ImplementMotorEncoders.cpp (from src/libYARP_dev/src/IMotorEncodersImpl.cpp)
+  src/libYARP_dev/src/ImplementPWMControl.cpp (from src/libYARP_dev/src/PWMControlImpl.cpp)
+  src/libYARP_dev/src/ImplementPidControl.cpp (from src/libYARP_dev/src/IPidControlImpl.cpp)
+  src/libYARP_dev/src/ImplementPositionControl.cpp (from src/libYARP_dev/src/IPositionControlImpl.cpp)
+  src/libYARP_dev/src/ImplementPositionDirect.cpp (from src/libYARP_dev/src/IPositionDirectImpl.cpp)
+  src/libYARP_dev/src/ImplementRemoteCalibrator.cpp (from src/libYARP_dev/src/RemoteCalibratorImpl.cpp)
+  src/libYARP_dev/src/ImplementRemoteVariables.cpp (from src/libYARP_dev/src/IRemoteVariablesImpl.cpp)
+  src/libYARP_dev/src/ImplementTorqueControl.cpp (from src/libYARP_dev/src/TorqueControlImpl.cpp)
+  src/libYARP_dev/src/ImplementVelocityControl.cpp (from src/libYARP_dev/src/IVelocityControlImpl.cpp)
+  src/libYARP_dev/src/ImplementVirtualAnalogSensor.cpp (from src/libYARP_dev/src/VirtualAnalogSensorImpl.cpp)
 * The following clsses have been refactored. They are no more template classes:
   ImplementControlCalibration
   ImplementAmplifierControl
