+++ src/libYARP_dev/include/yarp/dev/ControlBoardInterfacesImpl.h
@@ -26,12 +26,12 @@
 #include <yarp/dev/ImplementControlCalibration.h>
 #include <yarp/dev/ImplementAmplifierControl.h>
 #include <yarp/dev/ImplementEncoders.h>
-#include <yarp/dev/IPidControlImpl.h>
-#include <yarp/dev/IPositionControl2Impl.h>
-#include <yarp/dev/IVelocityControl2Impl.h>
-#include <yarp/dev/IControlLimitsImpl.h>
-#include <yarp/dev/IPositionDirectImpl.h>
-#include <yarp/dev/IInteractionModeImpl.h>
+#include <yarp/dev/ImplementPidControl.h>
+#include <yarp/dev/ImplementPositionControl2.h>
+#include <yarp/dev/ImplementVelocityControl2.h>
+#include <yarp/dev/ImplementControlLimits.h>
+#include <yarp/dev/ImplementPositionDirect.h>
+#include <yarp/dev/ImplementInteractionMode.h>
 
 // old interface implementation file
 #include <yarp/dev/ImplementControlBoardInterfaces.h>
