+++ src/libYARP_dev/include/yarp/dev/ImplementInteractionMode.h
