+++ src/libYARP_dev/include/yarp/dev/ImplementVirtualAnalogSensor.h
