+++ src/libYARP_dev/src/ImplementAmplifierControl.cpp
