+++ src/libYARP_dev/src/ImplementEncoders.cpp
